[AMD] Reorganize hip-related header files in sgl-kernel (#9320)
