+++ sgl-kernel/include/hip/hip_act_and_mul.cuh
